[#931] Add `ConnectionLimitManager` + Metrics.getNumberOfConnections() returns int.
Add the `ConnectionLimitManager` and use it for MQTT and AMQP protocol
adapters.
AMQP: If no connection limit is configured, the AMQP protocol adapter auto-
configures a limit based on the amount of memory and the configured
maximum session window size. If a limit is configured and it is above
the recommended value, a warning is printed.
The MQTT adapter uses now ConnectionLimitManager.

+++ service-base/src/main/java/org/eclipse/hono/service/metric/Metrics.java
@@ -51,7 +51,7 @@ public interface Metrics {
      *
      * @return total number of connections.
      */
-    long getNumberOfConnections();
+    int getNumberOfConnections();
 
     /**
      * Starts a new timer.
